api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.3.2
anthropic==0.7.1
faiss-cpu==1.7.4
pymupdf==1.23.5
python-dotenv==1.0.0
python-multipart==0.0.6
sentence-transformers==2.2.2
# azure-openai
